update tests and remove old operations
